Some rearranging and doc updates
