+++ examples/utils/migration-tools/src/migrator/makeMigratorEntryPointPiece.ts
@@ -9,13 +9,12 @@ import type { FluidObject } from "@fluidframework/core-interfaces";
 import type { IEntryPointPiece } from "../compositeRuntime/index.js";
 import { MigrationToolFactory, type IMigrationTool } from "../migrationTool/index.js";
 
-import type { IMigrator } from "./interfaces.js";
-import {
-	type ExportDataCallback,
-	type LoadSourceContainerCallback,
-	type MigrationCallback,
-	Migrator,
-} from "./migrator.js";
+import type {
+	ExportDataCallback,
+	LoadSourceContainerCallback,
+	MigrationCallback,
+} from "./interfaces.js";
+import { Migrator } from "./migrator.js";
 
 const migratorEntryPointPieceName = "getMigrator";
 
@@ -34,17 +33,6 @@ async function getDataStoreEntryPoint(
 	}
 
 	return entryPointHandle.get();
-}
-
-/**
- * The partial type of the entrypoint provided when makeMigratorEntryPointPiece is used.
- * @alpha
- */
-export interface IMigratorEntryPoint {
-	getMigrator: (
-		loadSourceContainerCallback: LoadSourceContainerCallback,
-		migrationCallback: MigrationCallback,
-	) => Promise<IMigrator>;
 }
 
 /**
